requirements.txt
beautifulsoup4==4.9.0
google==2.0.3
opuslib==3.0.1
protobuf==3.11.3
PyAudio==0.2.11
-e git+https://github.com/azlux/pymumble.git@37805ffba6148c4d082b70c0574b10de7c104a16#egg=pymumble
PyYAML==5.3.1
six==1.14.0
soupsieve==2.0
